ui: Refactors the code-mirror linting/restrict legacy tokens to HCL
